create Losses module
