+++ src/Flux.jl
@@ -35,13 +35,15 @@ include("onehot.jl")
 include("functor.jl")
 
 include("layers/stateless.jl")
-include("layers/losses.jl")
 include("layers/basic.jl")
 include("layers/conv.jl")
 include("layers/recurrent.jl")
 include("layers/normalise.jl")
 
 include("data/Data.jl")
+
+include("losses/Losses.jl")
+using .Losses
 
 include("deprecations.jl")
